LibWeb: Split RectStyleValue out of StyleValue.{h,cpp}
