+++ Userland/Libraries/LibWeb/CSS/StyleValue.h
@@ -657,28 +657,6 @@ private:
     } m_properties;
 };
 
-class RectStyleValue : public StyleValueWithDefaultOperators<RectStyleValue> {
-public:
-    static ValueComparingNonnullRefPtr<RectStyleValue> create(EdgeRect rect);
-    virtual ~RectStyleValue() override = default;
-
-    EdgeRect rect() const { return m_rect; }
-    virtual ErrorOr<String> to_string() const override;
-    virtual bool has_rect() const override { return true; }
-    virtual EdgeRect to_rect() const override { return m_rect; }
-
-    bool properties_equal(RectStyleValue const& other) const { return m_rect == other.m_rect; }
-
-private:
-    explicit RectStyleValue(EdgeRect rect)
-        : StyleValueWithDefaultOperators(Type::Rect)
-        , m_rect(rect)
-    {
-    }
-
-    EdgeRect m_rect;
-};
-
 }
 
 template<>
